Move profile assume role and sso to their own cred providers
* Wire them up to close their resources as well
